requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aioredis==2.0.1
aiosignal==1.4.0
alembic==1.17.2
altair==4.2.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
asyncio-throttle==1.0.2
asyncpg==0.29.0
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
aws_secretsmanager_caching==1.1.3
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
black==23.11.0
bleach==6.3.0
blinker==1.9.0
boto3==1.34.0
botocore==1.34.162
brotli==1.2.0
cachetools==6.2.4
category-encoders==2.6.1
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
ConfigArgParse==1.7.1
configobj==5.0.9
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==6.1.3
dpath==2.2.0
dulwich==0.25.0
dvc==3.65.0
dvc-data==3.17.0
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
ecdsa==0.19.1
entrypoints==0.4
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.104.1
fastjsonschema==2.21.2
feature-engine==1.6.0
filelock==3.20.1
flake8==6.1.0
Flask==3.1.2
Flask-BasicAuth==0.2.0
flask-cors==6.0.2
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
funcy==2.0
gevent==25.9.1
geventhttpclient==2.3.7
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
great-expectations==0.18.1
greenlet==3.3.0
gto==1.9.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
hydra-core==1.3.2
identify==2.6.15
idna==3.11
importlib_metadata==7.2.1
iniconfig==2.3.0
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kombu==5.6.1
lark==1.3.1
limits==5.6.0
locust==2.20.1
makefun==1.16.0
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.2.0
mlflow==2.9.2
msgpack==1.1.2
multidict==6.7.0
mypy==1.7.1
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.10.0
notebook==7.5.1
notebook_shim==0.2.4
numpy==1.24.3
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.15.0
onnxconverter-common==1.16.0
orjson==3.11.5
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pathspec==0.12.1
patsy==1.0.2
pillow==12.0.0
platformdirs==3.11.0
pluggy==1.6.0
pre-commit==3.5.0
prometheus-client==0.18.0
prometheus-fastapi-instrumentator==6.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.8
psutil==7.2.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==14.0.2
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydeck==0.9.1
pydot==4.0.1
pyflakes==3.1.0
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyparsing==3.3.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==4.0.0
python-multipart==0.0.6
pytz==2023.4
PyYAML==6.0.3
pyzmq==27.1.0
querystring-parser==1.2.4
redis==5.0.1
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
roundrobin==0.0.4
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.17.17
s3transfer==0.9.0
scikit-learn==1.3.2
scipy==1.11.1
scmrepo==3.6.0
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
skl2onnx==1.15.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.27.0
statsmodels==0.14.6
streamlit==1.52.2
structlog==23.2.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer==0.21.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.2
uvicorn==0.24.0
vine==5.1.0
virtualenv==20.35.4
voluptuous==0.16.0
waitress==2.1.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wrapt==2.0.1
xgboost==2.0.1
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.1.1
